# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Project Final",
    "version" : "v1",
    "title" : "ProjectFinal",
    "contact" : {
      "name" : "tesis",
      "email" : "leandro.cocchi@davinci.com"
    }
  },
  "host" : "*.*",
  "basePath" : "/",
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/issue/sprint/{issueId}" : {
      "post" : {
        "operationId" : "setIssueInActiveSprint",
        "parameters" : [ {
          "name" : "issueId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Issue"
            }
          }
        }
      }
    },
    "/issue/sprint/{sprint}" : {
      "get" : {
        "operationId" : "getAllBySprint",
        "parameters" : [ {
          "name" : "sprint",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Issue"
              }
            }
          }
        }
      }
    },
...
    "/priority" : {
      "get" : {
        "operationId" : "getAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Priority"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Priority"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Priority"
            }
          }
        }
      }
    },
    "/priority/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Priority"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Priority"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Priority"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/rol" : {
      "get" : {
        "operationId" : "getAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Rol"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Rol"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rol"
            }
          }
        }
      }
    },
    "/rol/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rol"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Rol"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rol"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/sprint" : {
      "get" : {
        "operationId" : "getAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Sprint"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SprintDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/sprint/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Sprint"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Sprint"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Sprint"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/state" : {
      "get" : {
        "operationId" : "getAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/State"
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/State"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/State"
            }
          }
        }
      }
    },
    "/state/{id}" : {
      "get" : {
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/State"
            }
          }
        }
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/State"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/State"
            }
          }
        }
      },
      "delete" : {
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Issue" : {
      "type" : "object",
      "required" : [ "backlog", "created", "enabled", "title", "updated" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "title" : {
          "type" : "string",
          "minLength" : 1
        },
        "avatar" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "priority" : {
          "type" : "string"
        },
        "label" : {
          "type" : "string"
        },
        "phase" : {
          "type" : "string"
        },
        "state" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "assignee" : {
          "type" : "string"
        },
        "reporter" : {
          "type" : "string"
        },
        "watcher" : {
          "type" : "integer",
          "format" : "int32"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updated" : {
          "type" : "string",
          "format" : "date-time"
        },
        "resolved" : {
          "type" : "string",
          "format" : "date-time"
        },
        "plannedStart" : {
          "type" : "string",
          "format" : "date-time"
        },
        "plannedEnd" : {
          "type" : "string",
          "format" : "date-time"
        },
        "estimated" : {
          "type" : "integer",
          "format" : "int32"
        },
        "remaining" : {
          "type" : "integer",
          "format" : "int32"
        },
        "sprint" : {
          "type" : "integer",
          "format" : "int32"
        },
        "backlog" : {
          "type" : "boolean"
        },
        "enabled" : {
          "type" : "boolean"
        }
      }
    },
...
    "Priority" : {
      "type" : "object",
      "required" : [ "created", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
    "Rol" : {
      "type" : "object",
      "required" : [ "created", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
    "Sprint" : {
      "type" : "object",
      "required" : [ "created", "dateFrom", "dateTo", "enabled", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "dateFrom" : {
          "type" : "string",
          "format" : "date-time"
        },
        "dateTo" : {
          "type" : "string",
          "format" : "date-time"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "enabled" : {
          "type" : "boolean"
        }
      }
    },
    "SprintDTO" : {
      "type" : "object",
      "required" : [ "date_from", "date_to", "name" ],
      "properties" : {
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "date_from" : {
          "type" : "integer",
          "format" : "int64"
        },
        "date_to" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "State" : {
      "type" : "object",
      "required" : [ "created", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1
        },
        "description" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
...
  }
}